Rename "Register" to Mobile Computing
