+++ app/src/androidTest/java/com/example/mobile_computing/ExampleInstrumentedTest.java
@@ -1,4 +1,4 @@
-package com.example.register_form;
+package com.example.mobile_computing;
 
 import android.content.Context;
