+++ app/src/main/AndroidManifest.xml
@@ -13,16 +13,16 @@
         android:theme="@style/Theme.Registor_Form"
         tools:targetApi="31">
         <activity
-            android:name="com.example.register_form.LogIn_Form"
+            android:name="com.example.mobile_computing.LogIn_Form"
             android:exported="false" />
         <activity
-            android:name="com.example.register_form.MainActivity"
+            android:name="com.example.mobile_computing.MainActivity"
             android:exported="false" />
         <activity
-            android:name="com.example.register_form.dataForm"
+            android:name="com.example.mobile_computing.dataForm"
             android:exported="false" />
         <activity
-            android:name="com.example.register_form.HelloPage"
+            android:name="com.example.mobile_computing.HelloPage"
             android:exported="true">
             <intent-filter>
                 <action android:name="android.intent.action.MAIN" />
