+++ app/src/main/java/com/example/mobile_computing/HelloPage.java
@@ -1,15 +1,12 @@
-package com.example.register_form;
+package com.example.mobile_computing;
 
 import android.os.Bundle;
 import android.view.View;
-import android.widget.Button;
 import android.widget.TextView;
 
 import androidx.activity.EdgeToEdge;
 import androidx.appcompat.app.AppCompatActivity;
-import androidx.core.graphics.Insets;
-import androidx.core.view.ViewCompat;
-import androidx.core.view.WindowInsetsCompat;
+
 import android.content.Intent;
 
 import com.example.registor_form.R;
