+++ app/src/main/java/com/example/mobile_computing/dataForm.java
@@ -1,4 +1,4 @@
-package com.example.register_form;
+package com.example.mobile_computing;
 
 import android.os.Bundle;
 import android.widget.TextView;
