+++ app/src/main/res/layout/activity_data_form.xml
@@ -5,7 +5,7 @@
     android:id="@+id/main"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context="com.example.register_form.dataForm"
+    tools:context="com.example.mobile_computing.dataForm"
     android:background="@drawable/bg"
     >
