+++ app/src/main/res/layout/activity_hello_page.xml
@@ -6,7 +6,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:background="@drawable/cat"
-    tools:context="com.example.register_form.HelloPage">
+    tools:context="com.example.mobile_computing.HelloPage">
 
     <TextView
         android:id="@+id/HelloID"
