+++ app/src/main/res/layout/activity_main.xml
@@ -6,7 +6,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:background="@drawable/bg"
-    tools:context="com.example.register_form.MainActivity">
+    tools:context="com.example.mobile_computing.MainActivity">
 
     <TextView
         android:layout_width="wrap_content"
