+++ app/src/main/res/values/strings.xml
@@ -1,5 +1,5 @@
 <resources>
-    <string name="app_name">Registor_Form</string>
+    <string name="app_name">Mobile Computing</string>
 
     <string-array name="provinces">
         <item>Abra</item>
