+++ app/src/test/java/com/example/mobile_computing/ExampleUnitTest.java
@@ -1,4 +1,4 @@
-package com.example.register_form;
+package com.example.mobile_computing;
 
 import org.junit.Test;
